edited Drive and created separate classes for autonomous and manual
