+++ InfiniteInsomnia/src/main/java/frc/robot/Robot.java
@@ -8,9 +8,6 @@
 package frc.robot;
 
 import edu.wpi.first.wpilibj.TimedRobot;
-import edu.wpi.first.wpilibj.smartdashboard.SendableChooser;
-import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
-
 /**
  * The VM is configured to automatically run this class, and to call the
  * functions corresponding to each mode, as described in the TimedRobot
@@ -20,7 +17,9 @@ import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
  */
 public class Robot extends TimedRobot {
   
-public Drive drive = new Drive();
+public ManualDrive drive = new ManualDrive();
+public AutonomousDrive autoDrive = new AutonomousDrive();
+
   /**
    * This function is run when the robot is first started up and should be
    * used for any initialization code.
@@ -55,7 +54,7 @@ public Drive drive = new Drive();
    */
   @Override
   public void autonomousInit() {
-  
+    autoDrive.autoStart();
   }
 
   /**
@@ -63,7 +62,7 @@ public Drive drive = new Drive();
    */
   @Override
   public void autonomousPeriodic() {
-    
+    autoDrive.autoDrive();
   }
 
   /**
